feat: create util to get couponbackgroundColor
