+++ src/components/mols/CouponActivated/index.tsx
@@ -4,7 +4,9 @@ import React from "react";
 import Image, { StaticImageData } from "next/image";
 import { AccordionInfo } from "@/components";
 import { getDateFormater } from "@/utils/getDateFormater";
+import { getBackgroundCouponColor } from "@/utils/getCouponBackgroundColor";
 import QRCode from "react-qr-code";
+import cx from "classnames";
 
 interface CouponProps {
   couponActivateCode: string;
@@ -32,7 +34,12 @@ const CouponActivated: React.FC<CouponProps> = ({
           {couponTitle}
         </h1>
         <div className="flex justify-end mb-1 mt-2 p-0.5 gap-4">
-          <span className="relative text-xl w-32 text-white bg-primary-main flex items-center justify-center px-6 py-3 rounded-full">
+          <span
+            className={cx(
+              getBackgroundCouponColor(Number(couponDiscount)),
+              "relative text-xl w-32 text-white flex items-center justify-center px-6 py-3 rounded-full",
+            )}
+          >
             {couponDiscount}%
             <span className="absolute -top-4 -left-4 w-10 h-10 rounded-full bg-gradient-to-r from-secondary-dark to-secondary-lighter"></span>
           </span>
