+++ src/components/mols/CouponCard/index.tsx
@@ -9,6 +9,7 @@ import CouponGreen from "@/images/easytolive/icons/coupongreen.svg";
 import ShoppingCartGreen from "@/images/easytolive/icons/shopping_cart_green.svg";
 import ClockCircleRed from "@/images/easytolive/icons/clock_circleRed.svg";
 import getExpirateTime from "@/utils/getExpirateTime";
+import { getBackgroundCouponColor } from "@/utils/getCouponBackgroundColor";
 import cx from "classnames";
 import { getDateFormater } from "@/utils/getDateFormater";
 
@@ -33,23 +34,6 @@ const CouponCard: React.FC<ICouponCardProps> = ({
   icon,
   setShowCouponModal,
 }) => {
-  function getBackgroundCouponColor(discount: number) {
-    const colorOptions = [
-      { value: 50, color: "bg-generic-alertRed" },
-      { value: 40, color: "bg-secondary-dark" },
-      { value: 30, color: "bg-secondary-main" },
-      { value: 25, color: "bg-secondary-light" },
-      { value: 15, color: "bg-primary-dark" },
-      { value: 5, color: "bg-primary-main" },
-    ];
-    const selectedColor = colorOptions.find(
-      ({ value }) => discount >= value,
-    ) || {
-      color: "bg-primary-light",
-    };
-    return selectedColor.color;
-  }
-
   return (
     <div
       onClick={() => setShowCouponModal && setShowCouponModal(true)}
