+++ src/utils/getCouponBackgroundColor.ts
@@ -1,0 +1,14 @@
+export function getBackgroundCouponColor(discount: number) {
+  const colorOptions = [
+    { value: 50, color: "bg-generic-alertRed" },
+    { value: 40, color: "bg-secondary-dark" },
+    { value: 30, color: "bg-secondary-main" },
+    { value: 25, color: "bg-secondary-light" },
+    { value: 15, color: "bg-primary-dark" },
+    { value: 5, color: "bg-primary-main" },
+  ];
+  const selectedColor = colorOptions.find(({ value }) => discount >= value) || {
+    color: "bg-primary-light",
+  };
+  return selectedColor.color;
+}
